Update PhyloGroup.sh
Updated with new code and parameters!

+++ PhyloGroup.sh
@@ -12,12 +12,14 @@ for arg in "$@"; do #This just sets it so that you can use the double --xxxx and
     "--output")   set -- "$@" "-y" ;;
     "--tree_genes")   set -- "$@" "-l" ;;
     "--tree_genomes")   set -- "$@" "-o" ;;
+    "--genomes_oi")   set -- "$@" "-t" ;;
+    "--genes_oi")   set -- "$@" "-r" ;;
     *)        set -- "$@" "$arg"
   esac
 done
 
 #This simply takes the arguement from the user and stores that in a variable.
-while getopts "p:h:y:l:o::" opt
+while getopts "p:h:y:l:o:t:r:" opt
 do
 	case "$opt" in
 		p)
@@ -35,6 +37,12 @@ do
 		o)
 			tree_genomes="${OPTARG}" 
 			;;
+		t)
+			genomes_oi="${OPTARG}" 
+			;;
+		r)
+			genes_oi="${OPTARG}" 
+			;;
 
 	esac
 done
@@ -46,10 +54,12 @@ genomeInterest=${genome_dir:-${working_directory}/Genomes_Interest}
 geneInterest=${gene_dir:-${working_directory}/Carriage_Genes}
 genesForTree=${tree_genes:-${working_directory}/Tree_Genes}
 genomesForTree=${tree_genomes:-${working_directory}/Tree_Genomes}
-plots=${output_dir:-${working_directory}/output}
+plots=${output_dir:-${working_directory}/tmp_output}
+genomes_oi=${genomes_oi:-"no"}
+genes_oi=${genes_oi:-"no"}
 
 #This checks whether input files are in the correct format or not
-python3 py/CheckFileFormat.py $genomesForTree $genesForTree $geneInterest $genomeInterest 
+python3 py/CheckFileFormat.py $genomesForTree $genesForTree $geneInterest $genomeInterest $genes_oi $genomes_oi 
 
 exit_status=$?  # store the exit status for later use
 
@@ -75,34 +85,41 @@ rm $plots/geneMatches.txt
 
 #First off, I'm removing parts of the names in the genomes we've been given - they aren't very good names but these are the only parts we can delete and still have unique ID's.
 #Changing to the directory containing genomes to be added in / checked.
-cd $genomeInterest
+
+if [ $genomes_oi == "yes" ]
+then
+	cd $genomeInterest
 
 #For example, changes EC0_06_19646_7#6.contigs_velvet.fa to EC0_06_19646_7.fasta. We only change extension to .fasta because the genome folder uses these extensions, which makes it easier in the script for the loops.
 #Side note, this isn't going to affect anything outside of the particular format that some genomes were given in. If the format changes to one more suitable, that's great.
-echo "Converting .fa to .fasta - this may not work if the file is already in .fasta - don't worry in this case. In the future, consider using .fasta from the beginning."
-for filename in *.fa; do 
-   mv "${filename}" "${filename%#*.fa}.fasta"
-   #filename=${filename%#*}
-   #mv ${filename} "${filename%.fa}.fasta"
-   echo "Changed $filename to have .fasta extension"
-done
+	echo "Converting .fa to .fasta - this may not work if the file is already in .fasta - don't worry in this case. In the future, consider using .fasta from the beginning."
+	for filename in *.fa; do 
+	   mv "${filename}" "${filename%#*.fa}.fasta"
+	   #filename=${filename%#*}
+	   #mv ${filename} "${filename%.fa}.fasta"
+	   echo "Changed $filename to have .fasta extension"
+	done
+fi
 
 #Change back to original directory
 cd $working_directory
 
-#Copying the genomes that were added in to the same genome folder for making tree.
-echo "Moving genomes added in to template genomes folder..."
-cp -a $genomeInterest/*.fasta $genomesForTree
+if [ $genomes_oi == "yes" ]
+then
+	#Copying the genomes that were added in to the same genome folder for making tree.
+	echo "Moving genomes added in to template genomes folder..."
+	cp -a $genomeInterest/*.fasta $genomesForTree
 
 #We make a list of the genomes we have added in -this is used to both remove the genomes from the default genome folder (in case new genomes are added, the old ones that were tested are removed) and for R plot.
-cd $genomeInterest 
-echo "Making list of genomes added in..."
-ls *.fasta | sed 's/.fasta//g' > $plots/genomes.added.txt
+	cd $genomeInterest 
+	echo "Making list of genomes added in..."
+	ls *.fasta | sed 's/.fasta//g' > $plots/genomes.added.txt
+fi
 
 #Change to place where genomes are stored and make a list of genome file names
 cd $genomesForTree
 echo "Making list of genomes used to make the tree..."
-ls *.fasta | sed 's/.fasta//g' > $plots/List.genomes.txt
+ls *.fasta | sed 's/.fasta//g' > $plots/listGenomes.txt
 
 #We make a selection of genes that are used to help separate the phylogroups of the genomes of interest, and we do this by a specific method in makeGenes.sh. 
 
@@ -113,7 +130,7 @@ cd $working_directory
 cd ClermonTyping-master/
 
 #We now run a ClermonTyping without Mash - which would also assign phylogroups, but isn't necessary because we don't use it.
-for x in `cat $plots/List.genomes.txt`
+for x in `cat $plots/listGenomes.txt`
 do
 	
 	makeblastdb -in $genomesForTree/$x.fasta -input_type fasta -out my_fasta -dbtype nucl #This makes blastdb iteratively for each genome
@@ -160,31 +177,27 @@ cd $working_directory
 
 echo "Removing strains assigned clade phylogroup"
 
-#Delete clades from phylogeny list
-#sed -i '/^I\b/d' $plots/phylogeny_list.txt 
-#sed -i '/^II\b/d' $plots/phylogeny_list.txt 
-#sed -i '/^III\b/d' $plots/phylogeny_list.txt 
-#sed -i '/^IV\b/d' $plots/phylogeny_list.txt 
-#sed -i '/^V\b/d' $plots/phylogeny_list.txt 
 #Generate new list of genomes to use
-#awk '{print $2}' $plots/phylogeny_list.txt | grep -f - $plots/List.genomes.txt > $plots/new_genome_list.txt
 python3 py/removeCladesFromList.py $working_directory
+
 #Copy new list over the original list
-cp $working_directory/new_genome_list.txt $plots/List.genomes.txt
+cp $working_directory/new_genome_list.txt $plots/listGenomes.txt
 cp $working_directory/new_phylogeny_list.txt $plots/group_list.txt
 cp $plots/group_list.txt $working_directory/group_list.txt
 
-
-#Removing the genomes of interest from the default genome list in Genomes
-echo "Removing genomes of interest from general genome file"
-cd $genomesForTree
-for f in `cat $plots/genomes.added.txt`
-do 
-    rm ${f}.fasta
-done
+if [ $genomes_oi == "yes" ]
+then
+	#Removing the genomes of interest from the default genome list in Genomes
+	echo "Removing genomes of interest from general genome file"
+	cd $genomesForTree
+	for f in `cat $plots/genomes.added.txt`
+	do 
+    		rm ${f}.fasta
+	done
+fi
 cd $working_directory
 
-mv $plots/List.genomes.txt $working_directory/List.genomes.txt
+mv $plots/listGenomes.txt $working_directory/listGenomes.txt
 #Removing temporary file
 rm -r $plots
